% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graficas_linea.R
\name{grafica_linea_nacional}
\alias{grafica_linea_nacional}
\title{Gráfico de línea de homicidios a nivel nacional}
\usage{
grafica_linea_nacional(df, tipo = NULL, por_sexo = FALSE)
}
\arguments{
\item{df}{Un data.frame con las columnas \code{ANIO_OCUR}, \code{ANIO_REGIS} y \code{SEXO}}

\item{tipo}{Tipo de año a usar: \code{"registro"} o \code{"ocurrencia"} (por defecto)}

\item{por_sexo}{Lógico. Si \code{TRUE}, muestra líneas por sexo. Si \code{FALSE}, solo la serie total.}
}
\value{
Un gráfico de línea (ggplot2)
}
\description{
Genera una gráfica de la evolución de homicidios por año, con opción de desagregar por sexo.
}
