% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabla_causa.R
\name{tabla_causa}
\alias{tabla_causa}
\title{Tabla de causas (REC_CAU) por año y nivel geográfico}
\usage{
tabla_causa(
  anio_tipo = NULL,
  desagregacion = NULL,
  geo_tipo = "ocurrencia",
  sexo = NULL,
  edad = NULL,
  periodo = NULL
)
}
\arguments{
\item{anio_tipo}{\code{"registro"}, \code{"ocurrencia"} o \code{NULL} (por defecto \code{"ocurrencia"}).}

\item{desagregacion}{\code{NULL}, \code{"entidad"} o \code{"municipio"}.}

\item{geo_tipo}{\code{"registro"} o \code{"ocurrencia"} (por defecto \code{"ocurrencia"}).}

\item{sexo}{\code{NULL}, \code{"hombres"} o \code{"mujeres"}.}

\item{edad}{Vector numérico opcional con una o más edades (entre 0 y 99).}

\item{periodo}{Vector numérico opcional con uno o más años (por ejemplo: \code{periodo = 2010}, \code{periodo = 2005:2010}, \code{periodo = c(2015, 2020)}).
Si se especifica, filtra los registros para que solo se incluyan los años indicados.}
}
\value{
Un \code{data.frame} con columnas \code{arma_fuego}, \code{arma_cortante}, \code{asfixia}, \code{otros}, \code{no_especificado},
y columnas geográficas si aplica.
}
\description{
Genera una tabla con el número de homicidios según categoría de causa
(variable \code{REC_CAU}), con opción de desagregación por año, entidad o municipio,
además de filtrado por sexo y grupo de edad.
}
\details{
Las causas consideradas son:
\itemize{
\item \code{arma_fuego} (1)
\item \code{arma_cortante} (2)
\item \code{asfixia} (3)
\item \code{otros} (4)
\item \code{no_especificado} (5)
}
}
