% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabla_dia.R
\name{tabla_dia}
\alias{tabla_dia}
\title{Tabla de homicidios por día de la semana (DIA_SEMANA) y año}
\usage{
tabla_dia(
  anio_tipo = NULL,
  desagregacion = NULL,
  geo_tipo = "ocurrencia",
  sexo = NULL,
  edad = NULL,
  periodo = NULL
)
}
\arguments{
\item{anio_tipo}{\code{"registro"}, \code{"ocurrencia"} o \code{NULL} (por defecto \code{"ocurrencia"}).}

\item{desagregacion}{\code{NULL}, \code{"entidad"} o \code{"municipio"}.}

\item{geo_tipo}{\code{"registro"} o \code{"ocurrencia"} (por defecto \code{"ocurrencia"}).}

\item{sexo}{\code{NULL}, \code{"hombres"}, \code{"mujeres"}.}

\item{edad}{Vector numérico opcional con una o más edades (entre 0 y 99). Por ejemplo:
\code{edad = 25}, \code{edad = c(15:19)} o \code{edad = c(0:17, 24, 50)}.}

\item{periodo}{Vector numérico opcional con uno o más años (por ejemplo: \code{periodo = 2010}, \code{periodo = 2005:2010}, \code{periodo = c(2015, 2020)}).
Si se especifica, filtra los registros para que solo se incluyan los años indicados.}
}
\value{
Un \code{data.frame} con columnas por año, claves geográficas (si aplica) y días de la semana (\code{Dia_1} a \code{Dia_7}).
}
\description{
Devuelve una tabla con la distribución de homicidios por día de la semana (\code{DIA_SEMANA}),
con opción de desagregación por año, entidad o municipio, filtrado por sexo y edad.
}
\details{
Los días están codificados del 1 al 7, donde:
\itemize{
\item \code{Dia_1}: Domingo
\item \code{Dia_2}: Lunes
\item \code{Dia_3}: Martes
\item \code{Dia_4}: Miércoles
\item \code{Dia_5}: Jueves
\item \code{Dia_6}: Viernes
\item \code{Dia_7}: Sábado
}
}
