% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabla_hora.R
\name{tabla_hora}
\alias{tabla_hora}
\title{Tabla de homicidios por hora del día (HORAS) y año}
\usage{
tabla_hora(
  anio_tipo = NULL,
  desagregacion = NULL,
  geo_tipo = "ocurrencia",
  sexo = NULL,
  edad = NULL,
  periodo = NULL
)
}
\arguments{
\item{anio_tipo}{\code{"registro"}, \code{"ocurrencia"} o \code{NULL} (por defecto \code{"ocurrencia"}).}

\item{desagregacion}{\code{NULL}, \code{"entidad"} o \code{"municipio"}.}

\item{geo_tipo}{\code{"registro"} o \code{"ocurrencia"} (por defecto \code{"ocurrencia"}).}

\item{sexo}{\code{NULL}, \code{"hombres"}, \code{"mujeres"}.}

\item{edad}{Vector numérico opcional con una o más edades (entre 0 y 99).
Por ejemplo: \code{edad = 25}, \code{edad = c(15:19)} o \code{edad = c(0:17, 24, 50)}.}

\item{periodo}{Vector numérico opcional con uno o más años (por ejemplo: \code{periodo = 2010}, \code{periodo = 2005:2010}, \code{periodo = c(2015, 2020)}).
Si se especifica, filtra los registros para que solo se incluyan los años indicados.}
}
\value{
Un data.frame con columnas \code{ANIO_XXX}, claves geográficas (si aplica), \code{hora_00} a \code{hora_23}, y \code{no_especificado}.
}
\description{
Devuelve una tabla con la distribución de homicidios según la hora del día (\code{HORAS}),
con opción de desagregación por año, entidad o municipio, y filtros por sexo y edad.
}
\details{
Las columnas generadas van de \code{hora_00} a \code{hora_23}, además de \code{no_especificado}.
}
