% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabla_ocupacion.R
\name{tabla_ocupacion}
\alias{tabla_ocupacion}
\title{Tabla de ocupación (CAT_OCU) por año y opcionalmente por entidad o municipio}
\usage{
tabla_ocupacion(
  df,
  anio_tipo = NULL,
  desagregacion = NULL,
  geo_tipo = "ocurrencia",
  sexo = NULL,
  edad = NULL
)
}
\arguments{
\item{df}{Un data.frame que contenga las columnas \code{CAT_OCU}, \code{ANIO_OCUR}, \code{ANIO_REGIS}, \code{SEXO} y claves geográficas.}

\item{anio_tipo}{\code{"registro"}, \code{"ocurrencia"} o \code{NULL} (usa \code{"ocurrencia"} por defecto).}

\item{desagregacion}{\code{NULL}, \code{"entidad"} o \code{"municipio"}.}

\item{geo_tipo}{\code{"registro"} o \code{"ocurrencia"} (por defecto \code{"ocurrencia"}).}

\item{sexo}{\code{NULL}, \code{"hombres"}, \code{"mujeres"}.}

\item{edad}{Vector numérico opcional con una o más edades (entre 0 y 99) que se desean filtrar.
Por ejemplo: \code{edad = 25}, \code{edad = c(15:19)} o \code{edad = c(0:17, 24, 50)}.
Si se deja como \code{NULL} (por defecto), se incluyen todas las edades.}
}
\value{
Un data.frame con columnas para cada categoría ocupacional: \code{agricultor}, \code{artesanos}, \code{empleados}, \code{no_trabajaba},
\code{comerciantes}, \code{operadores}, \code{directores}, \code{profesionales}, \code{no_especificado}, \code{no_aplica}, \code{seguridad}
}
\description{
Devuelve una tabla con la distribución de homicidios según categoría ocupacional,
con opción de desagregación por año, entidad o municipio y por sexo.
}
