% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabla_total.R
\name{tabla_total}
\alias{tabla_total}
\title{Tabla de homicidios totales por año, sexo y nivel geográfico}
\usage{
tabla_total(
  anio_tipo = NULL,
  desagregacion = NULL,
  geo_tipo = "ocurrencia",
  edad = NULL,
  periodo = NULL
)
}
\arguments{
\item{anio_tipo}{\code{"registro"}, \code{"ocurrencia"} o \code{NULL} (por defecto \code{"ocurrencia"}).}

\item{desagregacion}{\code{NULL}, \code{"entidad"} o \code{"municipio"}.}

\item{geo_tipo}{\code{"registro"} o \code{"ocurrencia"} (por defecto \code{"ocurrencia"}).}

\item{edad}{Vector numérico opcional con una o más edades (entre 0 y 99).}

\item{periodo}{Vector numérico opcional con uno o más años (por ejemplo: \code{periodo = 2010}, \code{periodo = 2005:2010}, \code{periodo = c(2015, 2020)}).
Si se especifica, filtra los registros para que solo se incluyan los años indicados.}
}
\value{
Un \code{data.frame} con totales por año, sexo (\code{Hombres}, \code{Mujeres}, \code{NE}, \code{Total}),
y, si se especifica, columnas geográficas como \code{NOM_ENT} o \code{NOMGEO}.
}
\description{
Genera una tabla con el número total de homicidios por año,
con opción de desagregación por sexo, entidad o municipio,
y filtrado por grupo de edad.
}
