% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabla_causa.R
\name{tabla_causa}
\alias{tabla_causa}
\title{Tabla de causas (REC_CAU) por año, opcionalmente por entidad o municipio}
\usage{
tabla_causa(
  df,
  anio_tipo = NULL,
  desagregacion = NULL,
  geo_tipo = "ocurrencia",
  sexo = NULL,
  edad = NULL
)
}
\arguments{
\item{df}{Un data.frame que contenga las columnas \code{REC_CAU}, \code{ANIO_OCUR}, \code{ANIO_REGIS}, \code{SEXO} y claves geográficas.}

\item{anio_tipo}{\code{"registro"}, \code{"ocurrencia"} o \code{NULL}.}

\item{desagregacion}{\code{NULL}, \code{"entidad"} o \code{"municipio"}.}

\item{geo_tipo}{\code{"registro"} o \code{"ocurrencia"} (por defecto \code{"ocurrencia"}).}

\item{sexo}{\code{NULL}, \code{"hombres"}, \code{"mujeres"}.}

\item{edad}{Vector numérico opcional con una o más edades (entre 0 y 99) que se desean filtrar.
Por ejemplo: \code{edad = 25}, \code{edad = c(15:19)} o \code{edad = c(0:17, 24, 50)}.
Si se deja como \code{NULL} (por defecto), se incluyen todas las edades.}
}
\value{
Un data.frame con columnas por causa: \code{arma_fuego}, \code{arma_cortante}, \code{asfixia}, \code{otros}, \code{no_especificado}
}
\description{
Devuelve una tabla con el número de homicidios por categoría de causa (REC_CAU),
con opción de desagregación por año, entidad o municipio y por sexo.
}
