% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabla_hora.R
\name{tabla_hora}
\alias{tabla_hora}
\title{Tabla de homicidios por hora del día (HORAS) y año}
\usage{
tabla_hora(
  df,
  anio_tipo = NULL,
  desagregacion = NULL,
  geo_tipo = "ocurrencia",
  sexo = NULL,
  edad = NULL
)
}
\arguments{
\item{df}{Un data.frame con las columnas \code{HORAS}, \code{ANIO_OCUR}, \code{ANIO_REGIS}, \code{SEXO} y claves geográficas.}

\item{anio_tipo}{\code{"registro"}, \code{"ocurrencia"} o \code{NULL} (por defecto \code{"ocurrencia"}).}

\item{desagregacion}{\code{NULL}, \code{"entidad"} o \code{"municipio"}.}

\item{geo_tipo}{\code{"registro"} o \code{"ocurrencia"} (por defecto \code{"ocurrencia"}).}

\item{sexo}{\code{NULL}, \code{"hombres"}, \code{"mujeres"}.}

\item{edad}{Vector numérico opcional con una o más edades (entre 0 y 99) que se desean filtrar.
Por ejemplo: \code{edad = 25}, \code{edad = c(15:19)} o \code{edad = c(0:17, 24, 50)}.
Si se deja como \code{NULL} (por defecto), se incluyen todas las edades.}
}
\value{
Un data.frame con columnas \code{hora_00} a \code{hora_23}, más \code{no_especificado}.
}
\description{
Devuelve una tabla con la distribución de homicidios según la hora del día (\code{HORAS}),
con opción de desagregación por año, entidad o municipio y por sexo.
}
