% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabla_mes.R
\name{tabla_mes}
\alias{tabla_mes}
\title{Tabla de homicidios por mes (MES) y año}
\usage{
tabla_mes(
  anio_tipo = NULL,
  desagregacion = NULL,
  geo_tipo = "ocurrencia",
  sexo = NULL,
  edad = NULL,
  periodo = NULL
)
}
\arguments{
\item{anio_tipo}{\code{"registro"}, \code{"ocurrencia"} o \code{NULL} (por defecto \code{"ocurrencia"}).}

\item{desagregacion}{\code{"entidad"}, \code{"municipio"} o \code{NULL}.}

\item{geo_tipo}{\code{"registro"} o \code{"ocurrencia"} (por defecto \code{"ocurrencia"}).}

\item{sexo}{\code{"hombres"}, \code{"mujeres"} o \code{NULL}.}

\item{edad}{Vector numérico opcional con una o más edades (entre 0 y 99) que se desean filtrar.}

\item{periodo}{Vector numérico opcional con uno o más años (por ejemplo: \code{periodo = 2010}, \code{periodo = 2005:2010}, \code{periodo = c(2015, 2020)}).
Si se especifica, filtra los registros para que solo se incluyan los años indicados.}
}
\value{
Un data.frame con columnas: \code{enero}, \code{febrero}, ..., \code{diciembre}
}
\description{
Devuelve una tabla con el número de homicidios por mes (\code{MES}) y año,
con opción de desagregación por entidad o municipio y por sexo.
}
