% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabla_ocupacion.R
\name{tabla_ocupacion}
\alias{tabla_ocupacion}
\title{Tabla de homicidios por ocupación (CAT_OCU), año y nivel geográfico}
\usage{
tabla_ocupacion(
  anio_tipo = NULL,
  desagregacion = NULL,
  geo_tipo = "ocurrencia",
  sexo = NULL,
  edad = NULL,
  periodo = NULL
)
}
\arguments{
\item{anio_tipo}{\code{"registro"}, \code{"ocurrencia"} o \code{NULL} (por defecto \code{"ocurrencia"}).
Especifica si el año a considerar es el de registro o el de ocurrencia.}

\item{desagregacion}{\code{NULL}, \code{"entidad"} o \code{"municipio"}.
Define el nivel geográfico para desagregar la información.}

\item{geo_tipo}{\code{"registro"} o \code{"ocurrencia"} (por defecto \code{"ocurrencia"}).
Indica si la desagregación geográfica corresponde al lugar de registro o ocurrencia.}

\item{sexo}{\code{NULL}, \code{"hombres"} o \code{"mujeres"}.
Si se especifica, filtra los homicidios por sexo.}

\item{edad}{Vector numérico opcional con una o más edades (entre 0 y 99).
Por ejemplo: \code{edad = 25}, \code{edad = c(15:19)} o \code{edad = c(0:17, 24, 50)}.
Si se deja como \code{NULL} (por defecto), se incluyen todas las edades.}

\item{periodo}{Vector numérico opcional con uno o más años (por ejemplo: \code{periodo = 2010}, \code{periodo = 2005:2010}, \code{periodo = c(2015, 2020)}).
Si se especifica, filtra los registros para que solo se incluyan los años indicados.}
}
\value{
Un \code{data.frame} con columnas para cada categoría ocupacional:
\code{agricultor}, \code{artesanos}, \code{empleados}, \code{no_trabajaba},
\code{comerciantes}, \code{operadores}, \code{directores}, \code{profesionales},
\code{seguridad}, \code{no_especificado}, \code{no_aplica}.
Puede incluir columnas geográficas como \code{NOM_ENT} o \code{NOMGEO} si se especifica \code{desagregacion}.
}
\description{
Genera una tabla con el número de homicidios según categoría ocupacional
(variable \code{CAT_OCU}), con opción de desagregación por año, entidad o municipio,
además de filtrado por sexo y grupo de edad.
}
