% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabla_total.R
\name{tabla_total}
\alias{tabla_total}
\title{Tabla de homicidios totales por año, sexo y nivel geográfico}
\usage{
tabla_total(
  df,
  anio_tipo = NULL,
  desagregacion = NULL,
  geo_tipo = "ocurrencia",
  edad = NULL
)
}
\arguments{
\item{df}{Un data.frame con las columnas de año, sexo, geografía y claves territoriales}

\item{anio_tipo}{\code{"registro"}, \code{"ocurrencia"} o \code{NULL} (usa \code{"ocurrencia"} por defecto)}

\item{desagregacion}{\code{NULL}, \code{"entidad"} o \code{"municipio"}}

\item{geo_tipo}{\code{"registro"} o \code{"ocurrencia"} (por defecto \code{"ocurrencia"})}

\item{edad}{Vector numérico opcional con una o más edades (entre 0 y 99) que se desean filtrar.
Por ejemplo: \code{edad = 25}, \code{edad = c(15:19)} o \code{edad = c(0:17, 24, 50)}.
Si se deja como \code{NULL} (por defecto), se incluyen todas las edades.}
}
\value{
Un data.frame con totales por año, sexo, y opcionalmente por entidad o municipio
}
\description{
Tabla de homicidios totales por año, sexo y nivel geográfico
}
